refactor: moved hooks folder out of auth directory
