--- client/src/pages/auth/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useAuthReducer';
-export * from './useAuthErrorhandler';
-export * from './useValidation';
